pyproject.toml
[build-system]
requires = ["setuptools>=68.2.2", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plexus"
version = "0.0.1"
description = "Plexus --> a SSL method for neuronal activity phenotypic profiling."
authors = [
    { name = "Parker Grosjean", email = "parker.grosjean@gmail.com" }
]
license = {text = "MIT"}
urls = { "Homepage" = "https://github.com/pgrosjean/plexus" }
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10"
dependencies = [
    "cellpose>=3.0.7",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.2",
    "tqdm>=4.66.1",
    "zarr>=2.16.1",
    "anndata==0.11.1",
    "hydra-core==1.3.2",
    "pandas==2.2.3",
    "pydantic==2.10.4",
    "scikit_learn==1.6.0",
    "scipy==1.14.1",
    "wandb>=0.19.1",
]

[project.entry-points.console_scripts]
plexus-train = "plexus.ssl_training.launch_training:main"
plexus-inference = "plexus.ssl_inference.inference:main"

[tool.setuptools.packages.find]
where = ["."]
